Support Double(or any PythonConvertible FloatingPoint types)

+++ Sources/TestPlot/main.swift
@@ -7,7 +7,7 @@ try? FileManager.default.removeItem(atPath: "/tmp/tensorboardx")
 let writer = SummaryWriter(logdir: "/tmp/tensorboardx")
 
 for i in 0..<100 {
-    writer.addScalar(tag: "scalar/scalar", scalar: Float(i), globalStep: i, date: Date(timeIntervalSince1970: Double(i)))
+    writer.addScalar(tag: "scalar/scalar", scalar: Double(i), globalStep: i, date: Date(timeIntervalSince1970: Double(i)))
 }
 
 for i in 0..<100 {
@@ -17,7 +17,7 @@ for i in 0..<100 {
 }
 
 for i in 0..<3 {
-    let image = Tensor<Float>(randomUniform: [128, 128, 3])
+    let image = Tensor<Double>(randomUniform: [128, 128, 3])
     writer.addImage(tag: "image", image: image, globalStep: i)
 }
 
@@ -35,7 +35,7 @@ for i in 0..<3 {
     writer.addText(tag: "text", text: "step: \(i)", globalStep: i)
 }
 
-let data = Tensor<Float>(randomNormal: [100, 10])
+let data = Tensor<Double>(randomNormal: [100, 10])
 let labels = (0..<100).map { _ in String(Int.random(in: 0..<10)) }
 //let labelImages = Tensor<Float>(randomNormal: [100, 3, 32, 32])
